[Feat] 공지사항 CRD 관리자 권한

+++ src/main/java/com/homeis/notice/controller/NoticeController.java
@@ -1,5 +1,6 @@
 package com.homeis.notice.controller;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -7,6 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -14,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.homeis.notice.dto.Notice;
 import com.homeis.notice.dto.NoticePaginationResponse;
 import com.homeis.notice.model.service.NoticeService;
+import com.homeis.util.JWTUtil;
 
 import lombok.RequiredArgsConstructor;
 
@@ -22,6 +25,7 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/notice")
 public class NoticeController {
 	private final NoticeService noticeService;
+	private final JWTUtil jwtUtil;
 
 	@GetMapping("/list")
 	public ResponseEntity<?> list(@RequestParam(value = "size", defaultValue = "10") int size,
@@ -38,19 +42,28 @@ public class NoticeController {
 	}
 
 	@PostMapping("/regist")
-	public ResponseEntity<?> regist(@RequestBody Notice notice) {
+	public ResponseEntity<?> regist(@RequestBody Notice notice, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenJob = jwtUtil.getJobFromToken(tokenHeader.substring(7));
+		if(!tokenJob.equals("관리자")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한 없음. 잘못된 접근.");
+		
 		int isSucceed = noticeService.insertNotice(notice);
 		return ResponseEntity.status(200).body(isSucceed);
 	}
 
 	@PutMapping("/update")
-	public ResponseEntity<?> update(@RequestBody Notice notice) {
+	public ResponseEntity<?> update(@RequestBody Notice notice, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenJob = jwtUtil.getJobFromToken(tokenHeader.substring(7));
+		if(!tokenJob.equals("관리자")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한 없음. 잘못된 접근.");
+		
 		int isSucceed = noticeService.updateNotice(notice);
 		return ResponseEntity.status(200).body(isSucceed);
 	}
 
 	@DeleteMapping("/delete")
-	public ResponseEntity<?> delete(@RequestBody Notice notice) {
+	public ResponseEntity<?> delete(@RequestBody Notice notice, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenJob = jwtUtil.getJobFromToken(tokenHeader.substring(7));
+		if(!tokenJob.equals("관리자")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한 없음. 잘못된 접근.");
+		
 		int isSucceed = noticeService.deleteNotice(notice);
 		return ResponseEntity.status(200).body(isSucceed);
 	}
